Relocate the result package into internal/signalio.
This is the last step in allowing two binaries two exist that can do
collection:
- a productionized worker for large scale collection
- a small scale CLI focused tool
